Added a Tools sub-directory and files
Created a tools directory for self-made modules. This way, the notebooks will be kept clearer and more attention can be given to the actual data analysis rather than the long plotting code cells. For now, only a plot_funcs module was created, additional ones may be required for the EIT work.