Fixed alternative method to get unique clusters.
There was a problem in the order of the indices given assigned to the unique remaining clusters which lead to dominant balances being assigned to the wrong clusters and a mismatch in the color coding. This was fixed by assigning the new cluster indices in the loop where the duplicates were identified.