Fixed mean update for different n_clusters/ n_features
The issue was not detected due to n_clusters and n_features being the same. But the means should have been updated by doing the dot product of the probabilities and samples for each component (not all together). So it was put in the loop to be done component wise.